requirements.txt
discord.py[voice]>=1.0.0
colorama
pyfiglet
python-valve
mcrcon
mcstatus
aiofiles
aioconsole
youtube-dl
uvloop